amended?

+++ include/progressBar.hpp
@@ -25,15 +25,15 @@ public:
 
     void display() const
     {
-        float progress = ticks / total_ticks;
-        unsigned int pos = (bar_width * progress);
+        float progress = (float) ticks / total_ticks;
+        int pos = (bar_width * progress);
 
         std::chrono::steady_clock::time_point now = std::chrono::steady_clock::now();
         auto time_elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(now-start_time).count();
 
         std::cout << "[";
 
-        for (unsigned int i = 0; i < bar_width; ++i) {
+        for (int i = 0; i < bar_width; ++i) {
             if (i < pos) std::cout << complete_char;
             else if (i == pos) std::cout << ">";
             else std::cout << incomplete_char;
